plugin/commands/subnet/subnet.go
package subnet

import (
	"github.com/IBM-Cloud/ibm-cloud-cli-sdk/plugin"
	"github.com/spf13/cobra"

	. "github.ibm.com/SoftLayer/softlayer-cli/plugin/i18n"
	"github.ibm.com/SoftLayer/softlayer-cli/plugin/metadata"
)

func SetupCobraCommands(sl *metadata.SoftlayerCommand) *cobra.Command {
	cobraCmd := &cobra.Command{
		Use:   "subnet",
		Short: T("Classic infrastructure Network subnets"),
		RunE:  nil,
	}
	cobraCmd.AddCommand(NewCancelCommand(sl).Command)
	cobraCmd.AddCommand(NewCreateCommand(sl).Command)
	cobraCmd.AddCommand(NewDetailCommand(sl).Command)
	cobraCmd.AddCommand(NewClearRouteCommand(sl).Command)
	cobraCmd.AddCommand(NewEditIpCommand(sl).Command)
	return cobraCmd
}

func SubnetNamespace() plugin.Namespace {
	return plugin.Namespace{
		ParentName:  "sl",
		Name:        "subnet",
		Description: T("Classic infrastructure Network subnets"),
	}
}
